fix: add mutex to hal_log to prevent USB CDC crash across cores
Core 1 panic (LoadProhibited at 0x00000000) was caused by concurrent
Serial.printf() calls from loop() on Core 1 and commTask on Core 0.
USB CDC on ESP32-S3 is NOT thread-safe.

Fix:
- Add dedicated s_log_mutex (binary semaphore) for Serial access
- Create mutex in hal_mutex_init() (called early in setup)
- Wrap Serial.printf in hal_log with mutex take/give
- Cast volatile float desiredSteerAngleDeg to double in loop() status

+++ ag-steer/steer-controller/src/main.cpp
@@ -270,7 +270,7 @@ void loop() {
                 g_nav.safety_ok ? "OK" : "KICK",
                 g_nav.work_switch ? "ON" : "OFF",
                 g_nav.steer_switch ? "ON" : "OFF",
-                desiredSteerAngleDeg,
+                (double)desiredSteerAngleDeg,
                 hal_net_is_connected() ? "UP" : "DOWN");
     }
